requirements.txt
py-cord
google-cloud-aiplatform
vertexai
python-dotenv